Eval code: Execution context: 1 1 in old version should be 2
In old SpiderMonkey implementations (Firefox), up to version 1.7, it was possible to pass a calling context as a second argument for eval function. Thus, if the context still exists, it is possible to influence private variables.
However, due to security reasons in modern engines it was fixed and is not significant anymore.